Add support for VirtualBox 7.1
