+++ test/unit/plugins/providers/virtualbox/base.rb
@@ -7,3 +7,4 @@ require_relative "../../../base"
 require_relative "support/shared/virtualbox_driver_version_4_x_examples"
 require_relative "support/shared/virtualbox_driver_version_5_x_examples"
 require_relative "support/shared/virtualbox_driver_version_6_x_examples"
+require_relative "support/shared/virtualbox_driver_version_7_x_examples"
