+++ website/content/docs/providers/virtualbox/index.mdx
@@ -12,9 +12,9 @@ Vagrant comes with support out of the box for [VirtualBox](https://www.virtualbo
 a free, cross-platform consumer virtualization product.
 
 The VirtualBox provider is compatible with VirtualBox versions 4.0.x, 4.1.x,
-4.2.x, 4.3.x, 5.0.x, 5.1.x, 5.2.x, 6.0.x, 6.1.x and 7.0.x. Other versions are unsupported and the provider
-will display an error message. Please note that beta and pre-release versions
-of VirtualBox are not supported and may not be well-behaved.
+4.2.x, 4.3.x, 5.0.x, 5.1.x, 5.2.x, 6.0.x, 6.1.x, 7.0.x, and 7.1.x. Other versions
+are unsupported and the provider will display an error message. Please note that
+beta and pre-release versions VirtualBox are not supported and may not be well-behaved.
 
 VirtualBox must be installed on its own prior to using the provider, or
 the provider will display an error message asking you to install it.
